task_005_fullstack_meeting_app/repository_after/meeting-scheduler/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
requests==2.32.3
python-dateutil==2.9.0.post0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2

# Optional Google Calendar
google-api-python-client==2.149.0
google-auth==2.35.0
